Package application and library
