+++ karel_robot/__init__.py
@@ -1,0 +1,7 @@
+"""This package provides the classes for robot Karel and his World.
+
+Importing this package does not start the screen and is meant
+for testing and code reuse by anyone interested.
+If you just want to write Karel programs, import `karel.run`.
+"""
+from .window import *
