+++ karel_robot/run/__init__.py
@@ -1,0 +1,14 @@
+"""Karel interactive program visualization.
+
+This package provides the following functions for writing programs.
+It is also used to interactively command Karel by keyboard.
+"""
+from .karel_run import *
+
+__all__ = ['move', 'turn_left', 'turn_right',
+           'pick_beeper', 'put_beeper', 'beeper_is_present',
+           'facing_north', 'facing_east', 'facing_south', 'facing_west',
+           'front_is_blocked', 'front_is_treasure',
+           'set_speed', 'set_karel_beepers', 'pause',
+           'interactive'
+           ]
