+++ langton.py
@@ -1,16 +1,16 @@
 #!/usr/bin/env python3
-from karel_run import *
+"""Make Karel move like Langton's ant.
+"""
+from karel_robot.run import *
 
 set_speed(100)
 
-while True:
-    # At a white square, turn 90° right, flip the color of the square, move forward one unit
-    if not beeper_is_present():
-        put_beeper()
-        turn_right()
-        move()
-    # At a black square, turn 90° left, flip the color of the square, move forward one unit
-    else:
-        pick_beeper()
-        turn_left()
-        move()
+while True:  # repeat
+    if beeper_is_present():  # At a black square
+        pick_beeper()  # flip the color of the square
+        turn_left()  # turn 90° left
+        move()  # move forward one unit
+    else:  # At a white square
+        put_beeper()  # flip the color of the square
+        turn_right()  # turn 90° right
+        move()  # move forward one unit
